worker/requirements.txt
pika>=1.3.0
redis>=4.0.0
pillow>=9.0.0

torch>=2.0.0
transformers>=4.36.0
accelerate>=0.20.0